pluto.chat/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.25
pymysql==1.1.0

# Security
slowapi==0.1.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7

# HTTP Client
httpx==0.26.0

# AI & Embeddings
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
openai==1.10.0

# Vector Database
pinecone-client==3.0.0

# File Processing
PyMuPDF==1.23.8
docx2txt==0.8

# Media Processing
cloudinary==1.38.0

# Email
aiosmtplib==3.0.1

# Text Processing
beautifulsoup4==4.12.3
lxml==5.1.0
